encoding/xml: update XML name character ranges to current XML spec.
The current/old character ranges are from the 1998 specification.
This patch updates them to the 2008 (Fifth Edition) specification
available at: https://www.w3.org/TR/xml/#sec-common-syn
As a consequence, the NameStartChar and NameChar sets allowed
in XML names are substantially liberalized following the XML spec,
and also vastly simplified to a much smaller number of ranges.
(This could benefit performance but that's not a goal of this change.)

Change-Id: I053b392276d0e53b0a193968620ef0484672fba6