cmd/compile: remove unused tuple ops when regenerating
Only the Select[01] op is potentially removed when regenerating
flags. This can potentially leave a dead tuple op.

When adding candidates for removal after flag regeneration, also
add the tuple op if the flag generator was a select op.

This seems a little more common now that PPC64 treats ANDCCconst
as a tuple op.

Change-Id: I0ed238e7cec83b5e3680c7c7f2749b23740ae518