cmd/compile: enable address folding for global symbols of shared library
Address folding is disabled in CL42172, the commit message of which
said that "In shared library, load/store of global is rewritten to
using GOT and temp register, which conflicts with the use of temp
register for assembling large offset.". But arm64 GOT doesn't use the
Go temp register, it uses the intra-procedure-call temporary registers
IP0(R16) and IP1(R17). and the address of global symbols are computed
by "ADRP+ADD(relocs)", so the restrictions on shared mode address
folding no longer apply, and testing on darwin/arm64 confirms this.

Removing this restriction is beneficial to further optimize the sequence
of "ADRP+ADD+LD/ST" to "ADRP+LD/ST(offset), so this CL removes it.

Change-Id: I0e9f7bc1723e0a027f32cf0ae2c41cd6df49defe