cmd/link: allocate new space at the physical end of the file on Linux
Workaround bug of kernel tmpfs: rare corruption of linker generated ELF
binaries. See discussion at issues 53804 and 53116.

On linux, we preallocate storage space from the physical end of the file by
specifying it explicitly, so the size we give it should be the increment,
instead of the total size.

[1]: https://lore.kernel.org/linux-mm/33b85d82.7764.1842e9ab207.Coremail.chenguoqic@163.com/
[2]: https://lore.kernel.org/linux-mm/20221101032248.819360-1-kernel@hev.cc/T/#u

Fixes #53116.
Related #53804.
Updates #37310.

Change-Id: If41b357f10b7505be91ca77fc5fa83c2ca20505a