cmd/compile/internal/inline,cmd/compile/internal/pgo: remove candHotNodeMap and candHotEdgeMap
We currently preprocess this data into global maps, but this data can
actually be queried directly from the graph quite easily. This
simplifies the code quite a bit. With the exception of inlinedCallSites
(to be removed), this limits use of pgo.Profile from package inline to a
few methods, which will enable unexporting much of the API in a future
CL.

One notable change here is that candHotEdgeMap aggregated by
caller+line, so if a single edge at that caller+line exceeded the
threshold, then all calls with that caller+line would be considered hot.
EdgeWeight matches caller+callee+line. The cases where that may
different:

- An indirect call, which thus can have different callees. An
  indirect call is not inlinable, so this has no impact.

- Completely distinct direct calls on the same line
  (e.g., `if foo() && bar() {`). It didn't really make sense to
  consider both hot just because they are on the same line anyway.

This is a change we likely want to make anyway for additional
specificity.

For #55022.

Change-Id: I46366f002e16119650d4698316e716684028de77