cmd/compile: add testing-flag guard to package-is-collected assert
On advice of the department of garbage collection, forcing
a garbage collection generally does not improve performance.
However, this is a good property to be able to test, and it
was added to the fma_test, which was turned into a not-short
test so that this won't escape notice (and the end-to-end
dependence on the flag was tested with an inserted failure
in testing the test).

Benchmarking reveals that it is profitable to avoid this GC,
with about 1.5% reduction in both user and wall time.

(48 p) https://perf.golang.org/search?q=upload:20221103.3
(12 p) https://perf.golang.org/search?q=upload:20221103.5

Change-Id: I4c4816d619735838a32388acf0cc5eb1cd5f0db5