cmd/compile/internal/pgo: use a slice for locations
Currently locations are stored in a map and looked up by ID from
the map. The IDs are usually small sequential integers (the Go
pprof CPU profiles are so). Using a slice is more efficient.

TODO: in theory the IDs are not guaranteed to be  small sequential
integers. Maybe handle that case more elegantly (we could ignore
weird profiles).

Change-Id: I9e20d5cebca3a5239636413e1bf2f0b273038031