cmd/dist: add curiously missing GOOS/GOARCH combinations
The cmd/dist cgo_test enumerates a large number of platforms in
various special cases. Some combinations are suspiciously absent. This
CL completes the combinations.

DO NOT REVIEW. We'll see what the trybots have to say. If some of
these should be omitted, we'll can hopefully document why. So far I've
confirmed that the following work: windows-arm64, android-*,
linux-ppc64 (no cgo). linux-loong64 fails on one of the newly enabled
tests, so I've masked that one.

Preparation for #37486.

Change-Id: I04c3348e4f422d74d51e714647ca3db379e6e919